Fix infinite loop that would occur when trying to use venv-update with some instances of brew python.
I was running into an issue where venv-update would go into an infinite loop of
execing itself.  This was because sys.path was the same as scratch.venv, but it
had a suffix of bin/...

To fix this, I made the comparison compare absolute paths instead of strings,
which seems like the semantically correct assertion.