Update pip-faster to support newer pip versions
Updated pip_faster.py to support recent versions of pip. Right now that
means pip >= 8.0.0.

The output of the pip command has changed quite a bit lately, so I had
to patch most of the tests. One notable difference is that now some
warnings and errors are going to stderr instead of stdout.